Remove PoolExtendState from engine module
Just define it in thinpool.rs and export it so that StratPool can use it
in tests.
