+++ src/engine/strat_engine/thinpool/mod.rs
@@ -8,4 +8,4 @@ mod thinids;
 #[allow(clippy::module_inception)]
 mod thinpool;
 
-pub use self::thinpool::{ThinPool, ThinPoolSizeParams, DATA_BLOCK_SIZE};
+pub use self::thinpool::{PoolExtendState, ThinPool, ThinPoolSizeParams, DATA_BLOCK_SIZE};
